# be-summit/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Summit API Documentation",
        "description": "API documentation for the BE Summit project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://be-summit.test",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/admin/basecamp": {
            "get": {
                "tags": [
                    "Basecamp (Admin)"
                ],
                "summary": "List all basecamps (Admin)",
                "operationId": "6a17b7aeed243b484bae0712a5e472a4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Basecamps fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Basecamps fetched successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BasecampResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Basecamp (Admin)"
                ],
                "summary": "Create a new basecamp (Admin)",
                "operationId": "15c7ebb30637b1e989d19f15fe4da797",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mitra_id",
                                    "jalur_id",
                                    "nama_basecamp",
                                    "latitude",
                                    "longitude",
                                    "jam_operasional"
                                ],
                                "properties": {
                                    "mitra_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "jalur_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nama_basecamp": {
                                        "type": "string",
                                        "example": "Basecamp Merbabu via Selo"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "-7.441234"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "110.421234"
                                    },
                                    "jam_operasional": {
                                        "type": "string",
                                        "example": "24 Jam"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Basecamp created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Basecamp created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BasecampResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/basecamp/{id}": {
            "get": {
                "tags": [
                    "Basecamp (Admin)"
                ],
                "summary": "Get details of a specific basecamp (Admin)",
                "operationId": "c84ae8582157ac942445efb52bd2542f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Basecamp ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Basecamp details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Basecamp details fetched successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BasecampResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Basecamp not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Basecamp (Admin)"
                ],
                "summary": "Update an existing basecamp (Admin)",
                "operationId": "a24bf0653a6da43d876bb631c34db469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Basecamp ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mitra_id",
                                    "jalur_id",
                                    "nama_basecamp",
                                    "latitude",
                                    "longitude",
                                    "jam_operasional"
                                ],
                                "properties": {
                                    "mitra_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "jalur_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nama_basecamp": {
                                        "type": "string",
                                        "example": "Basecamp Merbabu via Selo Updated"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "-7.441235"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "110.421235"
                                    },
                                    "jam_operasional": {
                                        "type": "string",
                                        "example": "07:00 - 22:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Basecamp updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Basecamp updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BasecampResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Basecamp not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Basecamp (Admin)"
                ],
                "summary": "Delete a basecamp (Admin)",
                "operationId": "00549515241163f821794f77ee3daf93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Basecamp ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Basecamp deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Basecamp deleted successfully."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Basecamp not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BasecampResource": {
                "title": "Basecamp Resource",
                "description": "Basecamp profile resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "mitra_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "jalur_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "nama_basecamp": {
                        "type": "string",
                        "example": "Basecamp Merbabu Selo"
                    },
                    "latitude": {
                        "type": "string",
                        "example": "-7.441234"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "110.421234"
                    },
                    "jam_operasional": {
                        "type": "string",
                        "example": "24 Jam"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-05T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-05T12:00:00+07:00"
                    },
                    "mitra": {
                        "$ref": "#/components/schemas/MitraResource"
                    },
                    "jalur": {
                        "$ref": "#/components/schemas/JalurPendakianResource"
                    },
                    "produks": {
                        "description": "List of products sold by this basecamp",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProdukResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "JalurPendakianResource": {
                "title": "Jalur Pendakian Resource",
                "description": "Trail profile resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "gunung_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_jalur": {
                        "type": "string",
                        "example": "Jalur Cibodas"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Jalur legendaris berbatu dengan pemandangan air panas."
                    },
                    "titik_awal_mdpl": {
                        "type": "string",
                        "example": "1300 MDPL"
                    },
                    "titik_akhir_mdpl": {
                        "type": "string",
                        "example": "2958 MDPL"
                    },
                    "waktu_tempuh": {
                        "type": "string",
                        "example": "7 Jam"
                    },
                    "status": {
                        "type": "string",
                        "example": "open",
                        "enum": [
                            "open",
                            "close"
                        ]
                    },
                    "panjang_jalur": {
                        "type": "string",
                        "example": "9.7 Km"
                    },
                    "tingkat_kesulitan": {
                        "type": "string",
                        "example": "sedang",
                        "enum": [
                            "mudah",
                            "sedang",
                            "sulit",
                            "ekstrem"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    }
                },
                "type": "object"
            },
...
            "MitraResource": {
                "title": "Mitra Resource",
                "description": "Partner profile resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "nama_pemilik": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "telepon": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Raya Summit No. 10"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Pemilik Basecamp Merbabu Indah",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif"
                    },
                    "npwp": {
                        "type": "string",
                        "example": "12.345.678.9-012.000",
                        "nullable": true
                    },
                    "nik": {
                        "type": "string",
                        "example": "3201234567890001"
                    },
                    "rekening_bank": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "nama_rekening": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "bank": {
                        "type": "string",
                        "example": "Bank BCA"
                    },
                    "ewallet": {
                        "type": "string",
                        "example": "081234567890",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-05T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-05T12:00:00+07:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User Resource",
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jhon Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jhon@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "pendaki"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-02T15:34:17+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-02T15:34:17+07:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}